Add getter accessor support for remove whole object. (see #268)
