+++ dev/column.h
@@ -9,6 +9,7 @@
 #include "tuple_helper.h"
 #include "default_value_extractor.h"
 #include "constraints.h"
+#include "column_traits.h"
 
 namespace sqlite_orm {
     
@@ -103,7 +104,7 @@ namespace sqlite_orm {
                 return res;
             }
         };
-        
+
         /**
          *  Column traits. Common case.
          */
@@ -115,144 +116,6 @@ namespace sqlite_orm {
          */
         template<class O, class T, class ...Op>
         struct is_column<column_t<O, T, Op...>> : public std::true_type {};
-        
-        /**
-         *  Getters aliases
-         */
-        template<class O, class T>
-        using getter_by_value_const = T (O::*)() const;
-        
-        template<class O, class T>
-        using getter_by_value = T (O::*)();
-        
-        template<class O, class T>
-        using getter_by_ref_const = T& (O::*)() const;
-        
-        template<class O, class T>
-        using getter_by_ref = T& (O::*)();
-        
-        template<class O, class T>
-        using getter_by_const_ref_const = const T& (O::*)() const;
-        
-        template<class O, class T>
-        using getter_by_const_ref = const T& (O::*)();
-        
-        /**
-         *  Setters aliases
-         */
-        template<class O, class T>
-        using setter_by_value = void (O::*)(T);
-        
-        template<class O, class T>
-        using setter_by_ref = void (O::*)(T&);
-        
-        template<class O, class T>
-        using setter_by_const_ref = void (O::*)(const T&);
-        
-        template<class T>
-        struct is_getter : std::false_type {};
-        
-        template<class O, class T>
-        struct is_getter<getter_by_value_const<O, T>> : std::true_type {};
-        
-        template<class O, class T>
-        struct is_getter<getter_by_value<O, T>> : std::true_type {};
-        
-        template<class O, class T>
-        struct is_getter<getter_by_ref_const<O, T>> : std::true_type {};
-        
-        template<class O, class T>
-        struct is_getter<getter_by_ref<O, T>> : std::true_type {};
-        
-        template<class O, class T>
-        struct is_getter<getter_by_const_ref_const<O, T>> : std::true_type {};
-        
-        template<class O, class T>
-        struct is_getter<getter_by_const_ref<O, T>> : std::true_type {};
-        
-        template<class T>
-        struct is_setter : std::false_type {};
-        
-        template<class O, class T>
-        struct is_setter<setter_by_value<O, T>> : std::true_type {};
-        
-        template<class O, class T>
-        struct is_setter<setter_by_ref<O, T>> : std::true_type {};
-        
-        template<class O, class T>
-        struct is_setter<setter_by_const_ref<O, T>> : std::true_type {};
-        
-        template<class T>
-        struct getter_traits;
-        
-        template<class O, class T>
-        struct getter_traits<getter_by_value_const<O, T>> {
-            using object_type = O;
-            using field_type = T;
-            
-            static constexpr const bool returns_lvalue = false;
-        };
-        
-        template<class O, class T>
-        struct getter_traits<getter_by_value<O, T>> {
-            using object_type = O;
-            using field_type = T;
-            
-            static constexpr const bool returns_lvalue = false;
-        };
-        
-        template<class O, class T>
-        struct getter_traits<getter_by_ref_const<O, T>> {
-            using object_type = O;
-            using field_type = T;
-            
-            static constexpr const bool returns_lvalue = true;
-        };
-        
-        template<class O, class T>
-        struct getter_traits<getter_by_ref<O, T>> {
-            using object_type = O;
-            using field_type = T;
-            
-            static constexpr const bool returns_lvalue = true;
-        };
-        
-        template<class O, class T>
-        struct getter_traits<getter_by_const_ref_const<O, T>> {
-            using object_type = O;
-            using field_type = T;
-            
-            static constexpr const bool returns_lvalue = true;
-        };
-        
-        template<class O, class T>
-        struct getter_traits<getter_by_const_ref<O, T>> {
-            using object_type = O;
-            using field_type = T;
-            
-            static constexpr const bool returns_lvalue = true;
-        };
-        
-        template<class T>
-        struct setter_traits;
-        
-        template<class O, class T>
-        struct setter_traits<setter_by_value<O, T>> {
-            using object_type = O;
-            using field_type = T;
-        };
-        
-        template<class O, class T>
-        struct setter_traits<setter_by_ref<O, T>> {
-            using object_type = O;
-            using field_type = T;
-        };
-        
-        template<class O, class T>
-        struct setter_traits<setter_by_const_ref<O, T>> {
-            using object_type = O;
-            using field_type = T;
-        };
     }
     
     /**
